refactor(cli): relocate canonical CLI to cli.py; __main__.py is a thin launcher
